fixed implementation, added javadoc and tests
- Fixed of by one error in i2osp2.
- Fixed return signature of os2ip method: int -> long.
- Fixed integer wrap around error in os2ip for input 0xFFFFFFFF.
- Reworked ByteEncoding implementation to define all edge cases.

+++ src/main/java/de/hsheilbronn/mi/utils/crypto/hpke/AeadId.java
@@ -124,7 +124,7 @@ public enum AeadId
 		if (value.length != 2)
 			throw new IllegalArgumentException("value.length != 2");
 
-		int aeadId = ByteEncoding.os2ip(value);
+		long aeadId = ByteEncoding.os2ip(value);
 
 		if (AES_128_GCM.id == aeadId)
 			return AES_128_GCM;
